# Optimize call path building