# Optimize Call Path Building

This PR significantly improves the performance of `generate_simple_call_paths` by reducing execution time from 39.27s to 9.70s - a 75% speedup. The optimizations maintain full functional equivalence, verified through input/output comparison testing.

## Changes
- Used `collections.deque` instead of lists for more efficient path buffer operations 
- Implemented a cross-reference caching system to avoid redundant lookups

## Performance Impact
- Before: 39.27 seconds
- After: 9.70 seconds

[_profile_before.txt](https://github.com/user-attachments/files/19427071/_profile_before.txt)
[_profile_after1.txt](https://github.com/user-attachments/files/19427074/_profile_after1.txt)

## Testing
Verified functional equivalence by comparing logged input/output pairs between the original and optimized versions. Zero differences were found in the output traces.


```
❯ diff <(cat ./_data/trace_before.txt |sort) <(cat ./_data/trace_after.txt|sort)

❯
```

Related links:
- https://github.com/rand-tech/xrefer/commit/1e826e76685b09fe57d3b547f3f857e16049e203#commitcomment-153130873
- https://github.com/mandiant/flare-gsoc/discussions/42