@rand-tech 2025-04-05
Ran the test as requested using the provided samples. Used the idapro module for headless execution (link to script in gist <https://gist.github.com/rand-tech/325709dbc97a46d9b05ffc656d5aa2e4#file-idacli-py>).



Both logs are attached for comparison. The optimized version maintains the same number of nodes discovered while delivering the performance improvements.
Let me know if you need anything else before merging!


[_c6b727d7cff517577db838db18ad17b46334d3c91c2e50893634e56cdc19e41f.rand_xrefer.txt](https://github.com/user-attachments/files/19616698/_c6b727d7cff517577db838db18ad17b46334d3c91c2e50893634e56cdc19e41f.rand_xrefer.txt)
[_c6b727d7cff517577db838db18ad17b46334d3c91c2e50893634e56cdc19e41f.xrefer.txt](https://github.com/user-attachments/files/19616699/_c6b727d7cff517577db838db18ad17b46334d3c91c2e50893634e56cdc19e41f.xrefer.txt)


[_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.rand_xrefer.txt](https://github.com/user-attachments/files/19616749/_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.rand_xrefer.txt)
[_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.xrefer.txt](https://github.com/user-attachments/files/19616750/_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.xrefer.txt)