@m-umairx 2025-04-05
@rand-tech  Awesome, thank you for pushing this optimization! I will merge this in a bit. 

Quick question though, were you able to get the vmray trace to load in that sample? I saw the non-supported parser error in your logs, but I am assuming those are because of the incomplete import of the xrefer modules in the headless script. Just want to make sure you were able to get it running to get an idea of how the trace ingestion looks like for your GSoC decision.