@rand-tech 2025-04-06
Yes, I was able to load it successfully. Thanks for checking! 
I adjusted the filename just to keep things consistent with the conditions mentioned here: https://github.com/mandiant/xrefer/pull/6#issuecomment-2757948596.


[_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.vmray.xrefer.txt](https://github.com/user-attachments/files/19619051/_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.vmray.xrefer.txt)
[_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.vmray.rand_xrefer.txt](https://github.com/user-attachments/files/19619050/_4bef733c2a02c1776591d77fe320016616ea10dd3f32c45fd786adbe7d014e1f.vmray.rand_xrefer.txt)